Make board , list and card

+++ actions/delete-user.action.ts
@@ -1,11 +1,16 @@
 "use server";
 
+import { z } from "zod";
 import { auth } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
 import { APIError } from "better-auth/api";
 import { revalidatePath } from "next/cache";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
+
+const DeleteUserSchema = z.object({
+  userId : z.string().min(1,"User id is required"),
+})
 
 export async function deleteUserAction({ userId }: { userId: string }) {
   const headersList = await headers();
@@ -20,6 +25,12 @@ export async function deleteUserAction({ userId }: { userId: string }) {
     throw new Error("Forbidden");
   }
 
+  const parsed = DeleteUserSchema.safeParse({userId});
+  if(!parsed.success){
+    return { success: false, error: "Invalid User ID" };
+  }
+
+
   try {
     await prisma.user.delete({
       where: {
@@ -28,10 +39,10 @@ export async function deleteUserAction({ userId }: { userId: string }) {
       },
     });
 
-    if (session.user.id === userId) {
-      await auth.api.signOut({ headers: headersList });
-      redirect("/auth/sign-in");
-    }
+    // if (session.user.id === userId) {
+    //   await auth.api.signOut({ headers: headersList });
+    //   redirect("/auth/sign-in");
+    // }
 
     revalidatePath("/dashboard/admin");
     return { success: true, error: null };
